requirements.txt
yt-dlp
Flask
gunicorn
requests
python-dotenv 